Automatically accept unknown SSH host keys for user jenkins
This commit configures SSH client for user jenkins to automatically
accept not-yet-known host keys. This avoids ssh client refusing
to connect because of not-yet-known host keys - this happens first time
a CI job is checking out stuff or similar and this causing job to fail.

Not great, but we do now want such failure nor we want to manually
connect to builder and accept the key each time we rebuild the
builder image.