Fix installing "build essential" for Trixie
At some point Trixie dropped `libdwarf-dev` package which is needed
for OMR build. This commit updates hook script to not install above
package is not available, making it work on Trixie again. If the package
is reintroduced later, it will be picked up automatically.

Not ideal but at least it does not break building Trixie images.